runner: move all related code into a module
