+++ src/lib.rs
@@ -86,18 +86,14 @@ pub mod db;
 pub mod dirs;
 pub mod docker;
 pub mod errors;
-mod ex_prepare;
-pub mod ex_run;
 pub mod experiments;
 mod git;
 mod native;
 pub mod report;
 pub mod results;
 mod run;
-pub mod run_graph;
+pub mod runner;
 pub mod server;
-mod tasks;
-mod toml_frobber;
 pub mod toolchain;
 pub mod tools;
