+++ src/runner/unstable_features.rs
@@ -1,0 +1,106 @@
+use config::Config;
+use errors::*;
+use experiments::Experiment;
+use results::TestResult;
+use std::collections::HashSet;
+use std::path::Path;
+use toolchain::Toolchain;
+use walkdir::{DirEntry, WalkDir};
+
+pub(super) fn find_unstable_features(
+    _config: &Config,
+    _ex: &Experiment,
+    source_path: &Path,
+    _toolchain: &Toolchain,
+    _quiet: bool,
+) -> Result<TestResult> {
+    let mut features = HashSet::new();
+
+    for entry in WalkDir::new(source_path)
+        .into_iter()
+        .filter_entry(|e| !is_hidden(e))
+    {
+        let entry = entry.chain_err(|| "walk dir")?;
+        if !entry
+            .file_name()
+            .to_str()
+            .map(|s| s.contains(".rs"))
+            .unwrap_or(false)
+        {
+            continue;
+        }
+        if !entry.file_type().is_file() {
+            continue;
+        }
+
+        let new_features = parse_features(entry.path())?;
+
+        for feature in new_features {
+            features.insert(feature);
+        }
+    }
+
+    let mut features: Vec<_> = features.into_iter().collect();
+    features.sort();
+    for feature in features {
+        info!("unstable-feature: {}", feature);
+    }
+
+    Ok(TestResult::TestPass)
+}
+
+fn parse_features(path: &Path) -> Result<Vec<String>> {
+    let mut features = Vec::new();
+    let contents = ::std::fs::read_to_string(path)?;
+    for (hash_idx, _) in contents.match_indices('#') {
+        let contents = &contents[hash_idx + 1..];
+        let contents = eat_token(Some(contents), "!").or_else(|| Some(contents));
+        let contents = eat_token(contents, "[");
+        let contents = eat_token(contents, "feature");
+        let new_features = parse_list(contents, "(", ")");
+        features.extend_from_slice(&new_features);
+    }
+
+    Ok(features)
+}
+
+fn is_hidden(entry: &DirEntry) -> bool {
+    entry
+        .file_name()
+        .to_str()
+        .map(|s| s.starts_with('.'))
+        .unwrap_or(false)
+}
+
+fn eat_token<'a>(s: Option<&'a str>, tok: &str) -> Option<&'a str> {
+    eat_whitespace(s).and_then(|s| {
+        if s.starts_with(tok) {
+            Some(&s[tok.len()..])
+        } else {
+            None
+        }
+    })
+}
+
+fn eat_whitespace(s: Option<&str>) -> Option<&str> {
+    s.and_then(|s| {
+        if let Some(i) = s.find(|c: char| !c.is_whitespace()) {
+            Some(&s[i..])
+        } else {
+            None
+        }
+    })
+}
+
+fn parse_list(s: Option<&str>, open: &str, close: &str) -> Vec<String> {
+    let s = eat_whitespace(s);
+    let s = eat_token(s, open);
+    if let Some(s) = s {
+        if let Some(i) = s.find(close) {
+            let s = &s[..i];
+            return s.split(',').map(|s| s.trim().to_string()).collect();
+        }
+    }
+
+    Vec::new()
+}
